Move posix.stat_result to os.stat_result on python 3
In python 3, posix.stat_result is a re-export of os.stat_result, while
in python 2 it was the reverse. Update typeshed to reflect this.

In some sense this arguably should not matter? It impacts mypyc, though
which perhaps takes typeshed a little too seriously.
